requirements.txt
python-docx
fastapi==0.114.1
openai==1.44.1
pdfplumber==0.11.4
python-dotenv==1.0.1
tiktoken==0.9.0
uvicorn==0.30.6
python-multipart==0.0.12
pymupdf==1.24.11
PyPDF2==3.0.1
pymongo==4.10.1
beanie
